+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityNetworkAuthenticationHandler.java
@@ -75,7 +75,7 @@ public class UserActivityNetworkAuthenticationHandler extends UserActivityBaseHa
 	}
 
 	@Override
-	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Integer> additionalActivityHistogram) {
+	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Double> additionalActivityHistogram) {
 		//do nothing
 	}
