+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityWorkingHoursHandler.java
@@ -84,7 +84,7 @@ public class UserActivityWorkingHoursHandler extends UserActivityBaseHandler {
 	}
 
 	@Override
-	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Integer> additionalActivityHistogram) {
+	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Double> additionalActivityHistogram) {
 
 	}
 }
