+++ fortscale/fortscale-webapp/src/main/java/fortscale/web/rest/ApiOrganizationActivityController.java
@@ -76,7 +76,7 @@ public class ApiOrganizationActivityController {
 
         //return the list as a list of OrganizationActivityData.LocationEntry (of the )
         return currentCountriesToCountDictionary.entrySet().stream()
-                .sorted((entrySet, entrySet2) -> -Integer.compare(entrySet.getValue(), entrySet2.getValue())) //sort them by count (reverse order - we want the bigger values in the beginning)
+                .sorted((entrySet, entrySet2) -> -Double.compare(entrySet.getValue(), entrySet2.getValue())) //sort them by count (reverse order - we want the bigger values in the beginning)
                 .limit(limit) //take only the top 'limit-number' of entries
                 .map(entry -> new OrganizationActivityData.LocationEntry(entry.getKey(), entry.getValue())) //create list
                 .collect(Collectors.toList());                                                              //of location entries
